Add regions modal test
